tests: Move common test actions to parent abstract TarTests class.
